Added more detailed error information when journal server is not found
We now differentiate between the error of not finding the device and not being able to ping the server. Not being able to find the device probably means that it is off or on a different wifi network. Being able to find the device, but not ping the server means that the server probably crashed.